Add files via upload
Road crashes are a major challenging issue for transportation authorities that has been an increasing
trend in recent years. It has also been documented as one of the world's major causes of death,
which requires an immediate response from all countries. Drowsy driving is one of the leading
causes of road accidents. Drowsy driving causes a significant proportion of road accidents, which
causes severe harm and death. Previous researchers give sufficient proof that drowsy driving has
an influence on road crashes. For driver drowsiness detection we use a convolution neural network
(CNN)  classifier for the given images set of Drowsy and Non Drowsy.